Adding a double casting ensures the result will be double

+++ bdd.workshop.calculator/Operator.cs
@@ -27,7 +27,7 @@ namespace bdd.workshop.calculator
 
         public static int Multiply(int a, int b) => a * b;
 
-        public static double Divide(int a, int b) => a / b;
+        public static double Divide(int a, int b) => ((double)a) / ((double)b);
 
         public static int Substract(int a, int b) => a - b;
